state_t to cpu_exec_state_t
The state_t class that lived inside the
execution/cpu/execute.cc was segfaulting
on destruction. This class is only used
in the execute function defined in
execution/cpu/execute.h.
The name was changed to cpu_exec_state_t
and the mysterious segfaulting went away.
Note that state_t exists in another source
files as a mechanism to keep track of state
there.

* usage outs passed in to applys_progress
* fixed bugs in cpu_exec_state_t constructor
  * set num remaining
  * notify inputs and check if input node