Rewrote graph_writer to use full_shape
The issue with graph writer before was (1) the mechanism
for viewing was not well defined with dimensions of size 1
and (2) the mechanism for viewing was not exposed at the right
level.

What changed with this commit is that a full_shape_t and full_dim_t
class were added and used.

* still need to update llama/modules.(h/cc)