Can't donate inputs with diff dtypes
Fixed bug where x = to_f32(x : f16) was occuring,
which is a problem since the output has twice as much
memory than the input.

Now the taskgraph execution engine asks the kernel manager
what inputs may be donatable from a kernel perspective, which
will not allow donating in the presence of dtype conversion.

And the kernel manager has a better idea of what inputs can
be donated, so more donations are actually happening now